Added a name classification result tab
Added a name classification result tab, mostly based on the code for the
word picture.
Restrict the examples linked to name classification with the original CQP
query to those texts containing a match for the query.
Configurable grouping of name categories.
Report if the selected corpora contain no name information or if no names
were found with the search. Skip name tables for groups with no names.