config.js: Fixed struct attribute selection lists
Fixed corpus text attributes with a dataset to have a selection list in the
extended search, as well as word attributes. Before this fix, they were
presented as text input fields because of not adding to struct_attributes
the new properties extended_template and controller that are now required
for select-type attributes. This bug had been present since updating
Kielipankki's Korp to Korp 4.

app/config.js:
  settings.fn.add_attr_extra_properties(): Add extra properties to
    struct_attributes in corpus settings, as well as to attributes.
  Corrected an error in the comment for settings.attr_extra_properties.