config la_murre: Abbreviated corpus ids: lam_*
LA murre: Abbreviated corpus ids to lam_ followed by a parish abbreviation,
to alleviate the problem of too long URLs for the Korp server. The old
la_murre_parish ids are retained as aliases that are rewritten to the new
ids. The parish name abbreviations follow the suggestions of Maria Vilkuna
based on the abbreviations used in the Dictionary of Finnish Dialects.

app/config.js:
  la_murre_grouping: Replaced the parish names as ids with abbreviations
    and added the old ids as separate items for aliases.
  Add the old ids as aliases for the new ones.
  locally_available_corpora: Replaced la_murre_.* with lam_.*.