Sidebar: Support url_opts.stringify_link
Sidebar: Support specifying in the corpus configuration a function
url_opts.stringify_link to generate the complete HTML code for the link
(the "a" element, both the URL and link text). This can be used, for
example, if an URL attribute may contain a list of URLs.

app/scripts/widgets.coffee:
  Sidebar.renderItem(): If attribute type is "url" and url_opts contains
    stringify_link, call it to produce the complete HTML code for the link.

app/scripts/bin/widgets.js:
  Compiled file.