Restrict search: prequery_within as defaultwitihin
Restricting search contexts: Fixed to pass the defaultwithin parameter to
the backend based on prequery_within if a search context restriction is
specified. Previously, it did not affect defaultwithin, causing incorrect
results with context restrictions.

However, if the search restriction context is broader than a sentence, the
map now shows incorrect (too many) results, since the query for the map
expands the main search to defaultwithin. To correct that, we might need to
be able to tell the backend which prequeries should be expanded and which
should not.

app/scripts/model.coffee:
  model.KWICProxy.makeRequest(), model.StatsProxy.makeParameters(): Call
    util.addPrequeryWithin().

app/scripts/util.coffee:
  Added util.addPrequeryWithin().

app/scripts/bin/model.js,
app/scripts/bin/util.js:
  Compiled files.