config Suomi24: Changed attribute name translations
Suomi24: Changed the translations of the attributes text_discussionarea and
text_subsections to conform better with the terminology used on the Suomi24
site, and added aliases for the old names, to be shown only in search (at
Krister's request). The aliases require new attributes, implemented at CWB
level as symbolic links to the actual attribute files.

app/config.js:
  sattrlist.s24_update: Added text_discussionarea_alias,
    text_subsections_alias, displayed only in search.

app/translations/corpora-fi.json:
  New translations for suomi24fi_sub, suomi24fi_sect; old translations as
    suomi24fi_sub_old, suomi24_fi_sect_old.