Localization: Default translations for missing ones
Added a facility to use a translation in another language or the
localization key if a translation for a key is missing in a language. The
translation is taken from the first language ("KEY" for localization key)
mentioned in the list settings.defaultTranslations.

Configured to prefer English and then the localization key.

app/scripts/util.coffee:
  Added util.addDefaultTranslations(): Add default translations to
    window.loc_data.

app/scripts/main.coffee:
  Call util.addDefaultTranslations() when preloading is done.

app/config.js:
  Define settings.defaultTranslations.

app/scripts/bin/main.js,
app/scripts/bin/util.js:
  Compiled files.