Config ScotsCorr: Word list: Preselect by regexp
ScotsCorr word selection list: When opening the list, preselect words by
matching the regular expression corresponding to the CQP expression. This
allows the user to specify "starts with", "ends with", "contains" or an
arbitrary regular expression and get all the matching words preselected.
("Is not" and "not regexp" select nothing.) This assumes that the regular
expressions in JavaScript and CQP are compatible. The selection list
returns a regular expression containing a list of fully expanded words as
alternatives.

app/modes/other_languages_mode.js:
  attrs.scotscorr_word: $scope.setSelected(): Set the preselected words
    based on a regular expression corresponding to the word condition.