Config: Defined (Fi)NER attributes
Defined attributes for (Fi)NER-tagged corpora, mostly as attributes of the
intra-sentence structure "ne" that are represented as token attributes in
Korp. The approach follows mostly that of Språkbanken but adds a couple of
new attributes, for example the name itself and a place name attribute that
can also be used for non-NER-tagged corpora.

The attributes are ne_name, ne_ex, ne_type, ne_subtype, ne_fulltype,
ne_placename, ne_placename_source. In addition, names have the positional
attributes nertag and nerbio.