Added restricted corpora access modal
When the user tries to access restricted corpora (e.g., via the URL
parameter "corpus") that are currently not accessible, display a modal
listing the non-accessible corpora and containing links to log in to Korp
(if the user is not yet logged in) and to the LBR for applying for access
rights.

This is a preliminary version for getting feedback.

To-do:
- Instead of physical Korp corpora, show logical corpora sharing the same
  URN and access rights.
- Show a separate LBR link for each logical corpus.