No JS error when no corpora selected at startup
Fixed a JavaScript error that occurred when no corpora were selected at
startup (or immediately after a Shibboleth login): view.updateReduceSelect
tried to apply $.grepObj to sentence attributes that were undefined. In
such cases, because of interrupted loading, the buttons in the corpus
selector did not show the texts, for example. (search.coffee)

Also try to make it more likely that the protected corpora unlocked by a
Shibboleth login are selected before calling view.initSearchOptions (which
in turn calls view.updateReduceSelect) by moving the call as late as
possible. However, this probably still does not ensure correct behaviour.