Add logical corpus information
Add logical corpus information (programmatically) to settings.corpora.*:
property logical_corpus points to a containing corpus folder or to the
corpus itself. If a corpus is in a corpus folder having the property
info.is_logical_corpus or info.urn, that folder is the logical corpus;
otherwise the logical corpus is the same as the physical one.