Restricted access modal: Stop click propagation
Restricted corpora access modal: Stop click propagation so that a click
outside the modal closes the modal but not the corpus selector if open.
However, a click on the Korp page header area still does not close the
modal but closes the corpus selector, whereas a click elsewhere works
correctly.