Config: Added lbr_id where needed
Config: Added property lbr_id to those corpora in which it is different
from metadata_urn: ParFin, ParRus, HCS2, Topling. In these cases, the LBR
application either covers multiple logical corpora (form Korp's point of
view) or the id refers to the metadata of an older, non-Korp version of the
corpus.