Compare: Fixed to show values if compiled on word
Compare searches result view: Show attribute values when the comparison is
compiled on word, which usually has no entry in the attributes property of
a corpus configuration. Before this fix, the values for "word" in the
compare result were shown as empty because of a JavaScript error. (This bug
was introduced in commit 7ed04d45 "Compare: Show values containing / (more)
correctly".)