Simple search: Fix pref/suff search with lemgrams
Simple search: Fixed prefix and suffix search to work when searching for a
lemgram from corpora without prefix and suffix attributes. Include
conditions on the prefix and suffix attributes in the search only if all
the selected corpora contain the attributes. Otherwise, construct a regexp
for the lemgram matching the prefix or suffix of the lemma and the POS.