Compress hash params for Shibboleth return URLs
Compress the hash parameters of the URL with zlib (pako) and Base64-encode
them for the return URL parameters of the Shibboleth login and logout URLs,
allowing longer parameter lists before exceeding the Apache URL length
limit. The encoded hash parameter is named "gz" and it is decompressed to
the original parameters when initializing Korp.