Remove unavailable corpora from config by default
Remove from configuration corpora which are unavailable to the backend, if
settings.handleUnavailableCorpora is defined and not "none" nor "fatal".
This makes it easier to use the frontend in a development or test
environment with only some corpora available. Folders that would be empty
after removing corpora are also removed. Unavailable corpora are also
removed from the list of preselected corpora.

The list of available corpora is first retrieved to
window.availableCorpora. To work properly, this requires a backend version
modified to return with "info" only the corpora that are actually
accessible (commits on 2017-12-13 by Jyrki Niemi).