Pass info to backend log: UI language, search mode
Pass information to the backend to be written to the backend log in the
query parameter "loginfo": currently, the Korp user interface language and
search mode.

In addition, a "loginfo" parameter was added to the makeRequest method of
several model proxies (KWICProxy, LemgramProxy, StatsProxy, GraphProxy,
NameClassificationProxy) to allow passing additional log info to the
backend. This is currently used only for the example KWIC, which adds
"ex_kwic" to loginfo to distinguish it from the usual KWIC.