KWIC download: Fixed to work without corpus_order
KWIC download: Fixed to work with only kwic in the result data, without
corpus_order or hits.

When the result data comes from the progress callback instead of the
success callback, it only has kwic. (Why?) Previously, this caused the KWIC
download button not to work (at least often) when trying to download the
first result after making a search (and perhaps also in some other cases).