Statistics: Make value grouping configurable
Made grouping (collapsing) values in the statistics result configurable:
replaced inline code removing a colon followed by anything with a call to
statisticsFomratting.makeGroupingValue (in config/statistics_config.js).
(This is intended to remove the trailing colon and digits in lemma, saldo
and lemgram values, to collapse together the values differing only in the
trailing digits.) The behaviour remains currently the same as before.