Other languages mode: Modify to work on Korp 9
app/modes/other_languages_mode.js:
- Modify to work on Korp 9:
  - Rename struct_attributes -> struct_attributes.
  - Change the values of context and within to be those used in Korp 9.

The configuration works in general: the public corpora are displayed in
the corpus chooser correctly and the currently unavailable non-public
corpora are dropped, but using individual corpora has not been tested
yet. Some configuration features are not yet supported, but they are
silently ignored.